requirements.txt
# ══════════════════════════════════════════════════════════════
# Autonomous Multi-Agent AI Organization — Python Dependencies
# Generated from active venv + actual code imports (March 2026)
# Install: pip install -r requirements.txt
# ══════════════════════════════════════════════════════════════

# ── Core Framework ─────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
starlette>=0.40.0

# ── LLM — Google Gemini (primary) ──────────────────────────────
google-genai>=0.5.0
google-auth>=2.28.0

# ── LLM — Optional fallback clients (all imported in base_agent.py) ──────────
# Install whichever providers you use. The system gracefully skips unavailable ones.
boto3>=1.34.0
openai>=1.12.0
anthropic>=0.18.0

# ── Kafka Event Bus ─────────────────────────────────────────────
# System falls back to in-memory mock if kafka-python is absent
kafka-python>=2.0.2

# ── Async & Networking ──────────────────────────────────────────
httpx>=0.27.0
websockets>=12.0
anyio>=4.3.0

# ── Database ────────────────────────────────────────────────────
sqlalchemy>=2.0.27
asyncpg>=0.29.0

# ── Code Execution & Sandboxing ─────────────────────────────────
docker>=7.0.0
gitpython>=3.1.41

# ── Static Analysis & Testing ───────────────────────────────────
bandit>=1.7.7
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# ── Code Quality ────────────────────────────────────────────────
black>=24.2.0
ruff>=0.3.0

# ── Observability ───────────────────────────────────────────────
structlog>=24.1.0
prometheus-client>=0.20.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-exporter-otlp>=1.23.0
langchain-core>=0.1.0
langsmith>=0.1.0

# ── Task Graph & DAG ────────────────────────────────────────────
networkx>=3.2.0

# ── Web Scraping (browser_tool) ─────────────────────────────────
beautifulsoup4>=4.12.0

# ── Utilities ───────────────────────────────────────────────────
redis>=5.0.1
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
tenacity>=8.2.0
rich>=13.7.0
PyYAML>=6.0.1
requests>=2.31.0
filelock>=3.13.0
textual>=0.52.0
testcontainers[kafka]
qdrant-client>=1.8.0
